Embed world mesh shader in bevy_world_mesh
